requirements.txt
asttokens==2.4.1
asyncio==3.4.3
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.0
debugpy==1.8.0
decorator==5.1.1
executing==2.0.1
greenlet==3.0.0
idna==3.4
ipykernel==6.26.0
ipython==8.17.2
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.5.0
lxml==4.9.3
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.26.2
packaging==23.2
pandas==2.1.3
parso==0.8.3
platformdirs==4.0.0
playwright==1.39.0
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pyaml==23.9.7
pyee==11.0.1
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
rauth==0.7.3
requests==2.31.0
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
wcwidth==0.2.10
yahoo-oauth==2.0
